fix CI blockers
